Fix JS to only change badge buttons when appropriate (#48)
* Fix JS to only change badge buttons when appropriate
Fixes https://github.com/magfest/magstock/issues/39. The basic problem here was that we had JS in regextra.html that assumed it would be on the preregistration form, but we had since added code to remove badge buttons after an attendee kicked in money. The JS then changed the new badge buttons, in some cases making it impossible to upgrade.

* Update JS to use jQuery's extend
Based on feedback from the PR, we now accomplish the same thing but with added jQuery.